Elite Bot Music Re-work (v1.0)
Re-wrote pretty much the entire bot to run on the latest discord.js v14 and discord-player v6. This is a major update to the open-source bot. Further code enhancements will also occur over the next couple of updates to make the bot even easier to configure, and bring new features.

As expected with a big change, there may be a chance of bugs, please report these as an issue on this repository!